Reorganized cleaned_data
