+++ cleaned_data/Preparatory Data/Gubernatorial by Precinct/gubernatorial 2022 scraper.py
