+++ cleaned_data/Preparatory Data/PollClosures.csv
@@ -1,0 +1,65 @@
+State,Closure Time,Synchronous,Confirmed
+AL,8:00p EST,TRUE,FALSE
+AK (AKST),12:00a EST,FALSE,TRUE
+AK (HAST),1:00a EST,FALSE,TRUE
+AZ,9:00p EST,TRUE,TRUE
+AK,8:30p EST,TRUE,FALSE
+CA,11:00p EST,TRUE,FALSE
+CO,9:00p EST,TRUE,TRUE
+CT,8:00p EST,TRUE,TRUE
+DE,8:00p EST,TRUE,TRUE
+DC,8:00p EST,TRUE,FALSE
+FL (EST),7:00p EST,FALSE,TRUE
+FL (CST),8:00p EST,FALSE,TRUE
+GA,7:00p EST,TRUE,TRUE
+HI,12:00a EST,TRUE,FALSE
+ID (MST),10:00p EST,FALSE,FALSE
+ID (PST),11:00p EST,FALSE,FALSE
+IL,8:00p EST,TRUE,FALSE
+IN (EST),6:00p EST,FALSE,TRUE
+IN (EST),7:00p EST,FALSE,TRUE
+IA,9:00p EST,TRUE,FALSE
+KS (CST),8:00p EST,FALSE,TRUE
+KS (MST),9:00p EST,FALSE,TRUE
+KY (EST),6:00p EST,FALSE,FALSE
+KY (CST),7:00p EST,FALSE,FALSE
+LA,9:00p EST,TRUE,FALSE
+ME,8:00p EST,TRUE,FALSE
+MD,8:00p EST,TRUE,TRUE
+MA,8:00p EST,TRUE,TRUE
+MI (EST),8:00p EST,FALSE,TRUE
+MI (CST),9:00p EST,FALSE,TRUE
+MN,9:00p EST,TRUE,TRUE
+MS,8:00p EST,TRUE,TRUE
+MO,8:00p EST,TRUE,TRUE
+MT,10:00p EST,TRUE,FALSE
+NE (CST),9:00p EST,TRUE,TRUE
+NE (MST),9:00p EST,TRUE,TRUE
+NV,10:00p EST,TRUE,TRUE
+NH,8:00p EST,TRUE,FALSE
+NJ,8:00p EST,TRUE,TRUE
+NM,9:00p EST,TRUE,FALSE
+NY,9:00p EST,TRUE,FALSE
+NC,7:30p EST,TRUE,FALSE
+ND (CST),10:00p EST,FALSE,FALSE
+ND (MST),11:00p EST,FALSE,FALSE
+OH,7:30p EST,TRUE,TRUE
+OK,8:00p EST,TRUE,FALSE
+OR (MST),10:00p EST,FALSE,FALSE
+OR (PST),11:00p EST,FALSE,FALSE
+PA,8:00p EST,TRUE,TRUE
+RI,8:00p EST,TRUE,TRUE
+SC,7:00p EST,TRUE,FALSE
+SD (CST),8:00p EST,FALSE,TRUE
+SD (MST),9:00p EST,FALSE,TRUE
+TN (EST),8:00p EST,TRUE,FALSE
+TN (CST),8:00p EST,TRUE,FALSE
+TX (CST),8:00p EST,FALSE,FALSE
+TX (MST),9:00p EST,FALSE,FALSE
+UT,10:00p EST,TRUE,FALSE
+VT,7:00p EST,TRUE,TRUE
+VA,7:00p EST,TRUE,TRUE
+WA,11:00p EST,TRUE,FALSE
+WV,7:30p EST,TRUE,TRUE
+WI,9:00p EST,TRUE,FALSE
+WY,9:00p EST,TRUE,TRUE
