+++ cleaned_data/Preparatory Data/Senate by Precinct/2018 Senate.csv
