+++ cleaned_data/Preparatory Data/Senate by Precinct/2022 Senate.csv
