+++ cleaned_data/Preparatory Data/Senate by Precinct/nyt_scraper.py
